+++ examples/demo_kafka.py
@@ -1,5 +1,5 @@
 '''
-Copyright (C) 2018-2022 Bryant Moscon - bmoscon@gmail.com
+Copyright (C) 2018-2023 Bryant Moscon - bmoscon@gmail.com
 
 Please see the LICENSE file for the terms and conditions
 associated with this software.
@@ -12,7 +12,12 @@ from cryptofeed.exchanges import Coinbase
 
 
 """
-You can run a consumer in the console with the following command
+The AIOKafkaProducer accepts configuration options passed as kwargs to the Kafka callback(s)
+either as individual kwargs, an unpacked dictionary `**config_dict`, or both, as in the example below.
+The full list of configuration parameters can be found at
+https://aiokafka.readthedocs.io/en/stable/api.html#aiokafka.AIOKafkaProducer
+
+You can run a Kafka consumer in the console with the following command
 (assuminng the defaults for the consumer group and bootstrap server)
 
 $ kafka-console-consumer --bootstrap-server 127.0.0.1:9092 --topic trades-COINBASE-BTC-USD
@@ -28,8 +33,14 @@ class CustomTradeKafka(TradeKafka):
 
 
 def main():
-    f = FeedHandler()
-    cbs = {TRADES: CustomTradeKafka(), L2_BOOK: BookKafka()}
+    common_kafka_config = {
+        'bootstrap_servers': '127.0.0.1:9092',
+        'acks': 1,
+        'request_timeout_ms': 10000,
+        'connections_max_idle_ms': 20000,
+    }
+    f = FeedHandler({'log': {'filename': 'feedhandler.log', 'level': 'INFO'}})
+    cbs = {TRADES: CustomTradeKafka(client_id='Coinbase Trades', **common_kafka_config), L2_BOOK: BookKafka(client_id='Coinbase Book', **common_kafka_config)}
 
     f.add_feed(Coinbase(max_depth=10, channels=[TRADES, L2_BOOK], symbols=['BTC-USD'], callbacks=cbs))